check PR titles instead of individual commit msgs
